added internal client filling to base listeners
The new events will be added to the internal representation of rooms
and so will new rooms. There was a problem with trying to find a way
to add new rooms before adding new events and eventually I decided
that it would be better to add these in the dispatchers rather than as
listeners. This ensures that they are executed in the right order ie
new rooms will always be added first before the events dispatcher can
possibly be dispatched.

The tests have been changed so that the package can be loaded before
the test is run (it used to run as the package was being built which
idk if that's bad, lquery seemed to do this which is where i stole the
idea from). Regardless I have decided it is bad because it makes it
harder to repeat tests when debugging something.