requests module changed
the requests module now creates errors and parses the json. This
happens because the `callback` thing was bad and didn't make much
sense.

This has reduced the complexity of all the code in src/main.lisp

requests now does this with a `guard-request` macro that is similar to
the old generate-generic-callback function, just without needing to
ninja its way around. This macro is then used inside
define-matrix-endpoint to wrap the request in the forms that will
parse then check the json for the `error` key matrix might return. In
future we could also check the status codes of the request if needed
but as matrix is good in this regard, that probably won't be necessary.