clearner generation
handle-request is now a function so a labels form doesn't appear in
all the definitions.

symbols are being printed with the origin package which is how it's
supposed to be but there's some things we don't want,

ie we don't want to have to load the package that generated the forms,
just a package that the user specifies with the things it needs like
handle-request and whatever is made by request-guard.