schema accepts import-from
we need to now make sure the generated file has it's dependancies
covered which means we need to change core.lisp and move the requests
into a seperate package. we will also probably want to not use :use
:requests :authetentication in the core package either so that the
symbols that are printed do not contain the autowrap package
designator. We could also change pretty printer not to use the package
designator at all.