generation symbols fixed
in order to generate the spec file and the requests api, load the
generate-matrix-requests system. The matrix-requests system will just
use the generated package and api file.

the functions have been changed to intern the symbols they create in
the target package to avoid being printed with their home package as
matrix-autowrap.

schema now needs a target package and the target directory is just a
pathname to a dir not a file as the package and api file need to be
put there.