new conditions
I don't really know why the old thing was doing search instead of
string=

I think it's because we thought the namespace of the error code would
change when the homeserver changed? but that's not what happens.

I think we should also rename the requests module something like api,
split it into client, media and conditions

Then cl-matrix will have to import each one. client and media
shouldn't be exporting the conditions for the conditions package.